requirements.txt
streamlit
pypdf2
langchain
langchain-openai
langchain-community
faiss-cpu
python-dotenv
